% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asin_sqrt.R
\name{asin_sqrt}
\alias{asin_sqrt}
\title{Arc-sine square root transform}
\usage{
asin_sqrt(x, ...)
}
\arguments{
\item{x}{vector of values to transform}
}
\value{
Values on the modified scale
}
\description{
Transform appropriate for proportion data
}
\details{
Follows pg. 69 of McCune & Grace (2002).  Appropriate for
    vegetation cover and other proportion data.  Forces data to 0-1
    scale if any original values outside this range.
}
\examples{
x  <- rnorm(99, 50, 20)
xt <- asin_sqrt(x)
plot(x, xt)

}
\references{
McCune, B., and J. B. Grace. 2002. Analysis of Ecological
    Communities. MjM Software, Gleneden Beach, Oregon, USA. 304 pp.
}
\seealso{
Other transformation functions: \code{\link{genlogtrans}},
  \code{\link{normalize}}
}
