% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basal_area.R
\name{basal_area}
\alias{basal_area}
\title{Tree basal area}
\usage{
basal_area(x, ...)
}
\arguments{
\item{x}{vector of tree DBHs in cm (metric)}
}
\value{
Numeric total basal area per unit:
}
\description{
Total tree basal area via simple geometric area of a circle.
}
\details{
Uses simple geometric area of a circle, summed across the vector.
}
\examples{
dbhs <- c(40, 23, 32.5, 80)   # vector of tree DBH values (cm)
ba <- basal_area(dbhs)         # m^2 per unit
ba

}
