% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_quantile.R
\name{cut_quantile}
\alias{cut_quantile}
\title{Cut by quantile}
\usage{
cut_quantile(x, n = 7, ordered = FALSE, labels = NULL, ...)
}
\arguments{
\item{x}{vector of numeric values to cut.}

\item{n}{number of groups desired.}

\item{ordered}{logical: should the result be an ordered factor?}

\item{labels}{if `labels = FALSE`, simple integer codes are returned
instead of a factor; see \code{\link[base]{cut}}.}

\item{...}{additional arguments passed to \code{\link[base]{cut}}.}
}
\value{
A factor is returned, unless `labels = FALSE` which results in an
    integer vector of level codes.
}
\description{
Cut a vector into roughly equal-sized groups, based on quantile
    breaks.
}
\details{
Useful to create balanced groups of _roughly_ equal size.
}
\examples{
set.seed(44)
x <- rnorm(99)   # vector of numeric values
(y <- cut_quantile(x, n = 3))
(z <- cut_quantile(x, n = 3, labels = FALSE))

}
