% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{mom}
\title{Describe}
\usage{
describe(x, na.rm = TRUE, digits = 2, type = 1, ...)

mom(x, na.rm = TRUE, digits = 2, type = 1, ...)
}
\arguments{
\item{x}{vector of values to be summarized}

\item{na.rm}{logical, defaults to TRUE}

\item{digits}{numeric, number of digits for rounding}

\item{type}{default=1, an integer between 1 and 3 selecting one of
the algorithms for computing skewness detailed in
\code{\link[e1071]{kurtosis}}}

\item{...}{further arguments passed to other methods}
}
\value{
For \code{describe}, a list with descriptive statistics: \itemize{
  \item mean = mean
  \item sd = standard deviation
  \item var  = variance
  \item sem = standard error of the mean
  \item cv = coefficient of variation
  \item n = count of real elements
  \item NAs = count of NA elements
  \item skw = skewness
  \item krt = kurtosis
  }\cr

For \code{mom}, a list with 4 moments:\itemize{
  \item mean = mean
  \item var  = variance
  \item skew = skewness
  \item kurt = kurtosis
 }
}
\description{
Descriptive summary statistics, by method-of-moments.
}
\details{
Method-of-moments summaries for first 4 moments.  \itemize{
  \item NEG kurtosis = flat distribution (platykurtic)
  \item POS kurtosis = peaked distribution (leptokurtic)
  \item ZERO kurtosis ~ the normal distribution (mesokurtic)
  \item NEG skewness indicates mean < median (left-skewed)
  \item POS skewness indicates mean > median (right-skewed)
  }
}
\examples{
x <- c(rnorm(99, 0.1), NA)
describe(x)
mom(x)

}
\seealso{
\code{\link[e1071]{kurtosis}},
\code{\link[e1071]{skewness}}
}
