% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicates}
\alias{duplicates}
\title{Determine all duplicate elements}
\usage{
duplicates(x, xret = FALSE, ...)
}
\arguments{
\item{x}{a vector or a data frame or an array or NULL.}

\item{xret}{logical, default \code{FALSE} returns a logical vector
indicating which rows are duplicates, \code{TRUE} returns the
duplicated elements.}

\item{...}{additional arguments passed to
\code{\link[base]{duplicated}}.}
}
\value{
Numeric vector indexing rows (sample units) in \code{spe1} which
    are nearest compositional neighbors of \code{spe2}. The order
    corresponds to row order in \code{spe2}.
}
\description{
Determines which elements of a vector or data frame
    are duplicates of other elements, and returns either a logical
    vector indicating which elements (rows) are duplicates, or the
    duplicated elements (rows) themselves.
}
\details{
Behavior nearly identical to \code{\link[base]{duplicated}}, except
    returns ALL duplicated elements, not simply the first.
}
\examples{
x <- c('a','b','c','c','c')
y <- data.frame(x, y=c(1,2,4,4,5))
z <- as.matrix(y)
duplicates(x) # duplicate_s_
duplicated(x) # compare with duplicate_d_
duplicates(y)
duplicated(y)
duplicates(z)
duplicated(z)

}
\seealso{
\code{\link[base]{duplicated}}, and
     https://stackoverflow.com/questions/7854433/.
}
