% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genlogtrans.R
\name{genlogtrans}
\alias{genlogtrans}
\title{Generalized log-transform}
\usage{
genlogtrans(x, ...)
}
\arguments{
\item{x}{vector of values to transform}
}
\value{
Values on a modified log10(x+1) scale
}
\description{
Log transform properly handling zero and very small values
}
\details{
Follows pg. 69 of McCune & Grace (2002).  Preserves original order
   of magnitudes while accounting for values = 0 and << 1.
}
\examples{
x  <- rnorm(99, 50, 20)
xt <- genlogtrans(x)
plot(x, xt)

}
\references{
McCune, B., and J. B. Grace. 2002. Analysis of Ecological
    Communities. MjM Software, Gleneden Beach, Oregon, USA. 304 pp.
}
\seealso{
Other transformation functions: \code{\link{asin_sqrt}},
  \code{\link{normalize}}
}
