% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Values NOT matching}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{y}{vector or NULL: the values to be matched against.}
}
\value{
A logical vector of the same length as \code{x}, indicating if a
     match was NOT located for each element of \code{x}: thus the
     values are TRUE or FALSE and never NA.
}
\description{
Returns a logical vector indicating if there is NOT a match for its
     left operand.
}
\details{
Inverse of the commonly used \code{\link[base]{\%in\%}}.
}
\examples{
a <- 1:5
b <- 4:9
a \%notin\% b

}
\seealso{
\code{\link[base]{\%in\%}}.
}
