% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indval.R
\name{indval}
\alias{indval}
\title{(Blocked) Indicator Species Analysis}
\usage{
indval(spe, grp, blk = NULL, cond_probs = TRUE, ...)
}
\arguments{
\item{spe}{array, where rows = SUs and cols = species.  Values may be numeric
abundances or presence/absence.}

\item{grp}{vector, describing group membership (required).}

\item{blk}{vector, describing block membership (optional).}

\item{cond_probs}{logical, should conditional probabilities be calculated for
when blocked structure exists? Default `TRUE`.}

\item{...}{further arguments currently ignored.}
}
\value{
List containing items:\cr
    - \code{cond_probs}:  conditional probabilities, when blocks exist.\cr
    - \code{A}: relative abundance component of IndVal per group.\cr
    - \code{B}: relative frequency component of IndVal per group.\cr
    - \code{IV}: IndVal, the indicator value per group.\cr
    - \code{IVmax}: maximum IndVal in the indicated group.\cr
    - \code{IVmax_dir}: signed max IndVal, when blocks exist.\cr
    - \code{maxgrp}: the group in which each species attains max IndVal.\cr
    - \code{sumIVmax}: sum of max IndVal across all species.\cr
    - \code{sig_expected}: number of significant indicator species expected
                                                               at random.\cr
    - \code{pval}: permutation p-values for each species.\cr
}
\description{
Blocked or unblocked versions of Dufrene-Legendre indicator species analysis.
   If blocked, then can also calculate conditional probabilities of
   gain/loss/stasis using Bayes rule. If blocked, then blocks are assumed to
   be sample units and groups are assumed to be the two sample events.
}
\details{
Unblocked version follows Dufrene and Legendre (1997).  Blocked indicator
   species analysis was first available in PC-ORD (McCune and Mefford 2017).
   For blocked designs, the current function can also calculate conditional
   probabilities of species gains, losses, and stasis across two sampling
   events.  Conditional probabilities use Bayes' rule.
}
\examples{
data(braun)
spe <- braun$spe[-1,]                         # force even number

# no group structure exists, so lets pretend we have revisit data
blk <- factor(rep(1:(NROW(spe)/2), times=2))  # 173 sites 'visited' twice
grp <- factor(rep(c(1,2), ea=NROW(spe)/2))    # 'visits' 1 or 2 for each site

# unblocked
iv  <- indval(spe, grp)

# blocked by site
ivb <- indval(spe, grp, blk, cond_probs=TRUE)

# see effects of blocking
plot(iv$IVmax, ivb$IVmax) ; abline(0,1)
table(unblocked = iv$maxgrp, blocked = ivb$maxgrp)

# given our haphazard group assignment, most species have poor indicator value
hist(ivb$IVmax, breaks=22, xlab='Max IndVal', main='')
iv$sig_expected   # num indicator species expected at random

# see how IndVal compares to conditional gain/loss probabilities blocked by site
a <- data.frame(ivb$cond_probs)
plot(a$prloss, a$prgain, col=ecole::colvec(ivb$IVmax_dir), pch=16)

}
\references{
Dufrene, M. and Legendre, P. 1997. Species assemblages and indicator
    species: the need for a flexible asymmetrical approach. Ecological
    Monographs 67(3):345-366.

McCune, B., and M. J. Mefford. 2017. PC-ORD. Multivariate Analysis
    of Ecological Data. Version 7. MjM Software Design, Gleneden
    Beach, OR.
}
\seealso{
\code{\link[labdsv]{indval}} or \code{\link[indicspecies]{strassoc}}
}
