% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makecwm.R
\name{makecwm}
\alias{makecwm}
\title{Community-weighted means (CWM)}
\usage{
makecwm(spp, trait, wa = TRUE, stand = c("minmax", "deviates", "none"), ...)
}
\arguments{
\item{spp}{array of species data, where rows=SUs, cols=species}

\item{trait}{array of traits data, where rows=species, cols=traits}

\item{wa}{logical, should weighted averaging step follow
standardization step?  If TRUE, then use abundance weighted
trait AVGS, if FALSE, then use abundance weighted trait TOTALS.}

\item{stand}{standardize each trait by its "minmax" (default),
"deviates", or "none"}
}
\value{
A community-weighted traits matrix (techically a data frame), where
   rows=SUs and cols=traits
}
\description{
Community-weighted means traits matrix (SU x traits)
}
\details{
Emulates PC-ORD; see also McCune and Grace (2002). Recommend not
changing default wa and stand settings, as these create the CWM
matrix that is most sensible for other analyses.
}
\examples{
# first generate test data
require(copula)
cc <- normalCopula(param=c(0.4,0.2,-0.8), dim=3, dispstr="un")
dd <- mvdc(copula=cc, margins=c("gamma", "beta", "t"),
              paramMargins=list(list(shape=2, scale=1),
                               list(shape1=2, shape2=2),
                                list(df=5)) )
spp <- rMvdc(999, dd)
spp <- abs(spp) # force positive
colnames(spp) <- c("sp1", "sp2", "sp1")
trt <- spp + rnorm(999, 1) * spp^2
trt <- t(trt)
# following McCune unpubl table:
# trts standardized by none, abund-weighted totals
c1  <- makecwm(spp, trt, wa=F, stand="none")
# traits standardized by minmax, abund-weighted totals
c2  <- makecwm(spp, trt, wa=F, stand="minmax")
# traits standardized by std deviates, abund-weighted totals
c3  <- makecwm(spp, trt, wa=F, stand="deviates")
# traits standardized by none, abund-weighted averages
c4  <- makecwm(spp, trt, wa=T, stand="none")
# traits standardized by minmax, abund-weighted averages
c5  <- makecwm(spp, trt, wa=T, stand="minmax")       # PREFERRED
# traits standardized by std deviates, abund-weighted averages
c6  <- makecwm(spp, trt, wa=T, stand="deviates")
# compare:
i <- 80
pairs(data.frame(c1[,i], c2[,i], c3[,i], c4[,i], c5[,i], c6[,i]),
     upper.panel=NULL, oma=c(2,2,0,0), cex=0.7, las=1)

}
\references{
McCune, B., and J. B. Grace. 2002. Analysis of Ecological
    Communities. MjM Software, Gleneden Beach, Oregon, USA. 304 pp.
}
