% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchval.R
\name{matchval}
\alias{matchval}
\title{Lookup a value}
\usage{
matchval(x, y, val = NULL, xid = NULL, yid = NULL)
}
\arguments{
\item{x}{data.frame containing values of interest.}

\item{y}{data.frame to be matched.}

\item{val}{character, name of column containing values in \code{x}.}

\item{xid}{character, name of ID column in \code{x}.}

\item{yid}{character, name of ID column in \code{y}.}
}
\value{
Values from \code{x$val}, ordered by \code{y$yid}.
}
\description{
Lookup a value based on matching IDs among two
    data.frames.
}
\details{
Similar to vlookup function found elsewhere.  Returns \code{NA} if
    no match is found.  If any IDs are duplicated, then only the
    first matching value is returned (just like
    \code{\link[base]{match}}).
}
\examples{
x <- data.frame(ident = letters[1:7], foo = rnorm(7))
y <- data.frame(ident = letters[5:9], bar = rnorm(5))
y$foo <- matchval(x, y, 'foo')
x
y

# be careful you use the intended column name!
colnames(y)[1] <- 'identifier'
matchval(x, y, 'foo')  # wrongly matches on existing 'foo' column
matchval(x, y, 'foo', 'ident', 'identifier')  # correct

}
\seealso{
\code{\link[base]{match}}.
}
