% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mom.R
\name{mom}
\alias{mom}
\title{Method of moments (MoM)}
\usage{
mom(x, na.rm = TRUE, digits = 2, type = 1, ...)
}
\arguments{
\item{x}{vector of values to be summarized}

\item{na.rm}{logical, defaults to TRUE}

\item{digits}{numeric, number of digits for rounding}

\item{type}{default=1, an integer between 1 and 3 selecting one of
the algorithms for computing skewness detailed in
\code{\link[e1071]{kurtosis}}}
}
\value{
List with 4 moments: \itemize{
  \item mean = mean
  \item var  = variance
  \item skew = skewness
  \item kurt = kurtosis
 }
}
\description{
Mean, variance, skewness, and kurtosis of a vector (distribution)
}
\details{
Method-of-moments summaries for first 4 moments. \itemize{
  \item NEG kurtosis = flat distribution (platykurtic)
  \item POS kurtosis = peaked distribution (leptokurtic)
  \item ZERO kurtosis ~ the normal distribution (mesokurtic)
  \item NEG skewness indicates mean < median (left-skewed)
  \item POS skewness indicates mean > median (right-skewed)
 }
}
\examples{
# require(e1071)
x <- rnorm(n, 0.1)
mom(x)

}
\seealso{
\code{\link[e1071]{kurtosis}},
\code{\link[e1071]{skewness}}
}
