% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_collapse.R
\name{mx_collapse}
\alias{mx_collapse}
\title{Collapse columns in a species matrix}
\usage{
mx_collapse(x, c1, c2, method = "max", ...)
}
\arguments{
\item{x}{array of species data, where rows = SUs and cols = species,
containing the two columns to be collapsed and synonymized}

\item{c1}{column name (to be retained)}

\item{c2}{column name (to be omitted)}

\item{method}{method to retain values, one of 'min', 'max' or 'mean'}

\item{...}{further arguments passed to other methods}
}
\value{
Modified dataframe or matrix, with one fewer column than the
original.
}
\description{
Collapse two matrix or data.frame columns into one, where new
values are the min, max, or mean.  Appropriate for synonymizing
species (columns) in a species abundance matrix.
}
\details{
Appropriate for synonymizing species (columns) in a species
abundance matrix. For example, if a species occurs twice in a plot,
simply take the highest, lowest, or mean of observed values.  The
name of the first column is retained, the second column is removed.
}
\examples{
# species abundance data
set.seed(1917)
spe <- data.frame(matrix(rnorm(30, 10, 50), 10, 3))
spe[spe < 0] <- 0
colnames(spe) <- c('Acer rubrum','Acer saccharum','Acer negundo')
spe
mx_collapse(spe, c1='Acer rubrum', c2='Acer negundo')
mx_collapse(spe, c1='Acer rubrum', c2='Acer negundo', method='mean')

}
