% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_dropzero.R
\name{mx_dropzero}
\alias{mx_dropzero}
\title{Drop zero-sum rows or columns}
\usage{
mx_dropzero(x, method = "col", silent = TRUE, ...)
}
\arguments{
\item{x}{array of species data, where rows = SUs and cols = species}

\item{method}{either one of \code{'col'} or \code{'row'}}

\item{silent}{logical, do not report number of rows or columns
dropped?}

\item{...}{further arguments passed to other methods}
}
\value{
Modified dataframe or matrix, with fewer rows or columns than the
    original.
}
\description{
Drop zero-sum rows or columns from a species abundance matrix,
    often needed before calculating dissimilarities.
}
\details{
Often useful before calculating certain dissimilarity measures.
}
\examples{
# species abundance data
set.seed(1917)
spe <- data.frame(matrix(rnorm(30, 10, 50), 10, 3))
spe[spe < 0] <- 0
colnames(spe) <- c('Acer rubrum','Acer saccharum','Acer negundo')
spe
mx_dropzero(spe) # returns unchanged abundance matrix or data.frame
spe[,2] <- 0
spe[4,] <- 0
mx_dropzero(spe, 'col') # second column removed
mx_dropzero(spe, 'row') # fourth row removed

}
