% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_valid.R
\name{mx_valid}
\alias{mx_valid}
\title{Test matrix validity}
\usage{
mx_valid(x, ...)
}
\arguments{
\item{x}{array of species data, where rows = SUs and cols = species}

\item{...}{further arguments passed to other methods}
}
\value{
Logical value, \code{TRUE} == valid, \code{FALSE} == not valid.
}
\description{
Affirm whether a species abundance matrix is valid (no NAs, no
    zero-sum rows, no zero-sum columns, all data 'connected').
}
\details{
Often useful before further analysis such as calculating certain
dissimilarity measures or during rejection sampling of simulated
data.  Data are strictly 'connected' when all sample units share
the same species pool (no disconnected species pools).
}
\examples{
# species abundance data
set.seed(1917)
spe <- data.frame(matrix(rnorm(30, 10, 50), 10, 3))
spe[spe < 0] <- 0
colnames(spe) <- c('Acer rubrum','Acer saccharum','Acer negundo')
spe
mx_valid(spe) # expect TRUE
spe[,2] <- 0
spe[4,] <- 0
mx_valid(spe) # expect FALSE

}
