% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x, ...)
}
\arguments{
\item{x}{vector of values to transform}
}
\value{
Values on the modified scale
}
\description{
Normalize (relativize) data to 0-1 scale
}
\details{
Relativizes data to 0-1 scale based on min/max values.
}
\examples{
x  <- rnorm(99, 50, 20)
xt <- normalize(x)
plot(x, xt)

}
\references{
McCune, B., and J. B. Grace. 2002. Analysis of Ecological
    Communities. MjM Software, Gleneden Beach, Oregon, USA. 304 pp.
}
\seealso{
Other transformation functions: \code{\link{asin_sqrt}},
  \code{\link{genlogtrans}}
}
