% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_corr.R
\name{ord_varexpl}
\alias{ord_varexpl}
\alias{ord_corr}
\title{Ordination correlations}
\usage{
ord_varexpl(object, diss, ...)

ord_corr(object, x, method = c("pearson", "kendall", "spearman"), ...)
}
\arguments{
\item{object}{ordination object.}

\item{diss}{matrix of original dissimilarities (should exactly match those
used in the ordination).}

\item{...}{further arguments passed to \code{stats::cor()}.}

\item{x}{matrix or data.frame of variables to correlate with the ordination.}

\item{method}{a character string indicating which correlation coefficient
to compute, one of "pearson" (default), "kendall", or "spearman".}
}
\value{
Matrix of correlations or variance explained, per ordination axis.
}
\description{
Correlations of ordination scores versus the original dissimilarity matrix,
the main species matrix, or a secondary matrix.
}
\details{
For \code{ord_varexpl}, the correlation is calculated between each ordination
axis versus the original dissimilarity matrix \code{diss}.  This can be
interpreted as the percentage of variance (in the original dissimilarity
matrix) explained.

For \code{ord_corr}, the correlation is calculated between each ordination
axis versus each column of the supplied matrix \code{x}.  This can be
interpreted as the strength of association between ordination scores and each
individual variable.
}
\examples{
data(smoky, package='ecole')
spe <- smoky$spe
env <- smoky$env
D   <- vegan::vegdist(spe, method='bray') # original dissimilarities
ord <- stats::cmdscale(D, k=5)            # ordination
ord_varexpl(ord, D)
ord_corr(ord, spe)
ord_corr(ord, env)

}
