% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_nms.R
\name{ord_nms}
\alias{ord_nms}
\title{NMS with randomization test and stepdown dimensionality selection}
\usage{
ord_nms(x, autopilot = c("quick", "medium", "slow"), method = "bray", ...)
}
\arguments{
\item{x}{array, where rows = SUs and cols = variables such as
environmental or traits values.}

\item{autopilot}{character, one of three possible thoroughness options.}

\item{method}{dissimilarity index, passed to `vegan::vegdist`.}

\item{...}{further arguments passed to `vegan::metaMDS`.}
}
\value{
List containing items returned by `vegan::metaMDS` for the selected model,
   appended with further items:\cr
    - \code{pval}:  permutation p-values for each dimensionality.\cr
    - \code{stress_real_vs_rnd}:  stress values, where each column =
    dimensionality, and each row = either the real-data stress value
    (first row only) or the randomized-data stress value (successive rows).\cr
}
\description{
Nonmetric multidimensional scaling with randomization test for stopping rules
    and stepdown dimensionality selection.
}
\details{
NMS with stopping rules for dimensionality based on PC-ORD
    (McCune and Mefford 2017).  This is time-consuming, use parallel
    processing where possible.
}
\examples{
data(smoky)
m   <- ord_nms(smoky$spe, 'quick')
# m   <- ord_nms(smoky$spe, 'medium', parallel=10)
# m   <- ord_nms(smoky$spe, 'slow', parallel=10)

}
\references{
McCune, B., and M. J. Mefford. 2017. PC-ORD. Multivariate Analysis
    of Ecological Data. Version 7. MjM Software Design, Gleneden
    Beach, OR.
}
\seealso{
\code{\link[vegan]{metaMDS}}
}
