% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_pca.R
\name{ord_pca}
\alias{ord_pca}
\title{PCA with randomization test}
\usage{
ord_pca(x, B = 999, ...)
}
\arguments{
\item{x}{array, where rows = SUs and cols = variables such as
environmental or traits values}

\item{B}{numeric, number of randomizations}

\item{...}{further arguments passed to other methods}
}
\value{
List containing items returned by `stats::prcomp`, appended with further
    items:\cr
    - \code{eig}:  eigenvalues.\cr
    - \code{varexpl}:  proportion of variance explained.\cr
    - \code{cumvar}:  cumulative variance explained.\cr
    - \code{V}:  matrix of correlations between variables and the PCA scores.\cr
    - \code{tab}:  table of randomization results by PCA axis.\cr
    - \code{stopping}:  number of suggested dimensions, based on different
        stopping rules.
}
\description{
Principal components analysis with randomization test for stopping rules.
}
\details{
PCA with stopping rules based on RndLambda, RndF, or broken-stick (Peres-Neto
   et al. 2005).  The current implementation hard-codes the cross-products
   matrix as a correlation matrix (i.e., data are scaled and centered).
}
\examples{
data(smoky)
pc <- ord_pca(smoky$env)                    # PCA with randomization
plot(pc$x, type='n', asp=1) ; text(pc$x, cex=0.7)  # plot first two axes
head(pc$tab)                                # better format than `summary()`
pc$stopping                                 # number of dimensions
pc$V                                        # V = correlations with PC axes

}
\references{
Peres-Neto, P.R., D.A. Jackson, and K.M. Somers. 2005. How many principal
   components? stopping rules for determining the number of non-trivial axes
   revisited. Computational Statistics and Data Analysis 49:974–997.
}
\seealso{
\code{\link[stats]{prcomp}}
}
