% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_multi.R
\name{outlier_multi}
\alias{outlier_multi}
\title{Multivariate outliers}
\usage{
outlier_multi(x, mult = 3, method = "bray", metric = "SD",
  plot = TRUE, perc = NULL, pick = NULL, ...)
}
\arguments{
\item{x}{array of species data}

\item{mult}{SD threshold for 'outliers' = the number of SD's away from
the mean \emph{distance} at which an outlier is defined;
default is 3 times the SD}

\item{method}{distance measure; follows \code{\link[vegan]{vegdist}}}

\item{metric}{defaults to 'SD'; only other alternative is 'MAD'}

\item{plot}{logical, should a histogram of distances be drawn?}

\item{perc}{numeric, returns outlier values beyond the specified percentile}

\item{pick}{numeric, returns this number of most extreme outliers}

\item{...}{further arguments passed to other methods}
}
\value{
A data frame with plot identifiers, calculated distances, and
    standard deviations from the mean distance, either with
    or without a histogram.
}
\description{
Find extreme values in ecological community data
}
\details{
Outlier analysis based on community distances. Emulates PC-ORD, but
    see also McCune and Grace (2002).
}
\examples{
require(vegan)
data(dune)
set.seed(19)
spe <- rbind(dune, runif(ncol(dune),0,1000)) # add an outlier row
outlier_multi(spe, mult=3, plot=FALSE)
outlier_multi(spe, mult=3)
outlier_multi(spe, perc=5)
outlier_multi(spe, pick=5)

}
\references{
McCune, B., and J. B. Grace. 2002. Analysis of Ecological
    Communities. MjM Software, Gleneden Beach, Oregon, USA. 304 pp.
}
\seealso{
\code{\link{outlier_uni}} for univariate analog.

Other outlier functions: \code{\link{outlier_uni}}
}
\concept{outlier functions}
