% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_uni.R
\name{outlier_uni}
\alias{outlier_uni}
\title{Univariate outliers}
\usage{
outlier_uni(x, mult = 1.5, ...)
}
\arguments{
\item{x}{vector of numeric values to query}

\item{mult}{IQR threshold for 'outliers'; default is 1.5 times the IQR}

\item{...}{further arguments passed to other methods}
}
\value{
List of with 4 elements: threshold values (\code{lwrlim},
    \code{uprlim}) and indices of values beyond those
    thresholds (\code{lower}, \code{upper})
}
\description{
Find extreme values in a vector
}
\details{
Given a vector, finds indices of values exceeding an IQR threshold.
}
\examples{
set.seed(19)
x <- data.frame(x=seq(1:10),y=c(rnorm(6,1),rnorm(2,-9),rnorm(2,9)))
(outs <- outlier_uni(x$yy))
plot(x)
points(x[outs$lower,], col=2, pch=19)
points(x[outs$upper,], col=4, pch=19)

}
\references{
davidrroberts.wordpress.com/2015/08/26/r-code-outlier-function/
}
\seealso{
\code{\link{outlier_multi}} for multivariate analog.

Other outlier functions: 
\code{\link{outlier_multi}()}
}
\concept{outlier functions}
