% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plo.R
\name{plo}
\alias{plo}
\title{Simple scatterplot with useful defaults}
\usage{
plo(x, y = NULL, col = NULL, cex = 0.7, pch = 16, ...)
}
\arguments{
\item{x, y}{vectors of x- and y-axis values.}

\item{col}{vectors of point colors.}

\item{cex}{numeric vector for scaling point size.}

\item{pch}{a vector of plotting characters or symbols: see
\code{\link[graphics]{points}}.}

\item{...}{further arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot object.
}
\description{
Scatterplot with sensible default settings that permit
    overplotting many points.
}
\details{
Useful default settings for a simple scatterplot, closed circles
    with transparent black color that permits overplotting.
    Follows \code{graphics::plot.default}.
}
\examples{
set.seed(23)
n <- 50
plo(runif(n), rnorm(n))
n <- 500
plo(runif(n), rnorm(n))
n <- 5000
plo(runif(n), rnorm(n))

}
