% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of species abundance matrix}
\usage{
plot_heatmap(x, xord = TRUE, yord = FALSE, logbase = FALSE,
  labcex = 0.7, col, xexp = 1, yexp = 1, ...)
}
\arguments{
\item{x}{array of species data, where rows = SUs and cols = species}

\item{xord, yord}{logical, should rows, columns be ordered by mean?}

\item{logbase}{numeric, logarithm base to increase plot contrast}

\item{labcex}{numeric, label expansion factor; defaults to 0.7}

\item{col}{vector defining a color palette}

\item{xexp, yexp}{expansion factor making room for x,y labels}

\item{...}{further arguments passed to other methods}
}
\value{
None, plots to device.
}
\description{
Plot site x species matrix with heatmap colors and proper labels.
}
\details{
Plots a heatmap to quickly visualize species abundances/occurences.
}
\examples{
data(smoky)
z <- smoky$spe
plot_heatmap(z, yord=FALSE, yexp=1.7)
plot_heatmap(z, yord=FALSE, yexp=1.7, logbase=10)
plot_heatmap(z, yord=FALSE, yexp=1.7, logbase=10, asp=1)
plot_heatmap(z, xord=FALSE, yord=FALSE, yexp=1.7, logbase=10,
     asp=1)

}
