% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joy.R
\name{plot_joy}
\alias{plot_joy}
\title{Joy plots}
\usage{
plot_joy(m, fcol = rgb(0, 0, 0, 0.5), lcol = "black", lwid = 1,
  xinc = 0, xexp = 0.2, yinc = ncol(m), yexp = 1.2, ypad = 1,
  scaled = TRUE, pts = NULL, labcex = 0.5, ...)
}
\arguments{
\item{m}{matrix or data.frame, where each column vector will have
its own density polygon}

\item{fcol}{polygon fill color(s)}

\item{lcol}{polygon line color}

\item{lwid}{polygon line width}

\item{xinc}{horizontal offset for each polygon (0 = aligned)}

\item{xexp}{x-axis expansion factor for entire plot}

\item{yinc}{vertical offset for ea polygon}

\item{yexp}{vertical expansion factor for each polygon}

\item{ypad}{vertical padding for entire plot}

\item{scaled}{logical, should polygons be scaled to maximum?}

\item{pts}{optional vector of points to plot onto each row}

\item{labcex}{expansion factor for y-axis text labels}

\item{...}{further arguments passed to other methods}
}
\value{
None, plots to device.
}
\description{
Plot small multiples of many overlapping density distributions,
    also known as a joy plot or stacked distributions.
}
\details{
Plots one density polygon for each column in the data.frame or
matrix, resulting in potentially overlapping small multiples.
Polygons can be scaled vertically to their maximum, or else allowed
to vary freely.  Setting \code{yexp = 1}, \code{ypad = 1} and
\code{scaled = T} gives scaled polygons that fill each row with no
overlap.
}
\examples{
set.seed(23)
N <- 99
x <- rnorm(N,10,1)
y <- x^1.2 + rnorm(N,0,1)
z <- x^1.4 + rnorm(N,0,1)
m <- data.frame(x,y,z)
# allow overlap
plot_joy(m, ypad=1.05, labcex=1, xlab='Gradient X')
# forbid overlap
plot_joy(m, yexp = 1, ypad=1, labcex=1, xlab='Gradient X')
# color vector
plot_joy(m, fcol=c("#FF000080", "#A020F080", "#0000FF80"),
         yexp = 1, ypad=1, labcex=1, xlab='Gradient X')

}
