% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\alias{plot_ortho}
\alias{plot_loess}
\alias{plot_kernel}
\alias{plot_spline}
\alias{plot_gam}
\title{Plotting several kinds of regression line}
\usage{
plot_lm(
  x,
  y,
  col = "#00000040",
  lcol = "#FF0000BF",
  cex = 0.8,
  pch = 16,
  xlab = NULL,
  ylab = NULL,
  args.func = list(),
  ...
)

plot_ortho(
  x,
  y,
  col = "#00000040",
  lcol = "#FF0000BF",
  cex = 0.8,
  pch = 16,
  xlab = NULL,
  ylab = NULL,
  args.func = list(),
  ...
)

plot_loess(
  x,
  y,
  col = "#00000040",
  lcol = "#FF0000BF",
  cex = 0.8,
  pch = 16,
  xlab = NULL,
  ylab = NULL,
  args.func = list(),
  ...
)

plot_kernel(
  x,
  y,
  col = "#00000040",
  lcol = "#FF0000BF",
  cex = 0.8,
  pch = 16,
  xlab = NULL,
  ylab = NULL,
  args.func = list(),
  ...
)

plot_spline(
  x,
  y,
  col = "#00000040",
  lcol = "#FF0000BF",
  cex = 0.8,
  pch = 16,
  xlab = NULL,
  ylab = NULL,
  args.func = list(),
  ...
)

plot_gam(
  x,
  y,
  col = "#00000040",
  lcol = "#FF0000BF",
  cex = 0.8,
  pch = 16,
  xlab = NULL,
  ylab = NULL,
  args.func = list(),
  ...
)
}
\arguments{
\item{x, y}{vectors of x- and y-axis values.}

\item{col}{vector of point colors.}

\item{lcol}{vector of line colors.}

\item{cex}{numeric vector for scaling point size.}

\item{pch}{a vector of plotting characters or symbols: see
\code{\link[graphics]{points}}.}

\item{xlab, ylab}{character strings for axis labels.}

\item{args.func}{further arguments passed to the regression function.}

\item{...}{further arguments passed to \code{plot}.}
}
\value{
A plot object.
}
\description{
Plot a regression line for linear, orthogonal, local-polynomial,
   kernel, cubic smoothing spline, or GAM regressions.
}
\details{
Plots a scatterplot with a regression line, useful for quick
    exploratory analysis.
}
\examples{
set.seed(23)
n <- 55
x <- sin(1:n/pi)
y <- runif(n, 0, 0.75) + x
x[12] <- NA  # handles NA
set_par(6)
plot_lm(x, y, main='Linear')
plot_ortho(x, y, main='Orthogonal') # not identical to linear!
plot_loess(x, y,  args.func=list(span=0.2), main='Loess')
plot_kernel(x, y, args.func=list(bandwidth=0.2), main='Kernel')
plot_spline(x, y, args.func=list(df=11), main='Spline')
plot_gam(x, y, args.func=list(gamma=0.5), main='GAM')

}
\seealso{
\code{\link[stats]{lm}},
   \code{\link[stats]{prcomp}},
   \code{\link[stats]{loess}},
   \code{\link[stats]{ksmooth}},
   \code{\link[stats]{smooth.spline}}, and
   \code{\link[mgcv]{gam}} for the underlying functions.
}
