% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marg.R
\name{plot_marg}
\alias{plot_marg}
\title{Scatterplot with marginals}
\usage{
plot_marg(d, x, y, over, xlab, ylab, ovlab, marg = "dens", alpha = 1, ...)
}
\arguments{
\item{d}{array of data, where rows = SUs and cols = variables of interest}

\item{x}{character string for which column in \code{d} to place on x axis}

\item{y}{character string for which column in \code{d} to place on y axis}

\item{over}{character string for which column in \code{d} to overlay}

\item{xlab}{character string for label on x axis}

\item{ylab}{character string for label on y axis}

\item{ovlab}{character string for label on overlay axis when \code{marg='ortho'}}

\item{marg}{either \code{'hist'} for histograms,
   \code{'dens'} for density plots,
or \code{'ortho'} for orthogonal fitlines}

\item{alpha}{alpha level of fill colors}

\item{...}{further arguments passed to other methods}
}
\value{
A plot object
}
\description{
Scatterplot with marginal histograms, densities or orthogonal fit
    lines, useful for visualizing gradients on ordinations.
}
\examples{
set.seed(23)
d <- data.frame(xx=rnorm(100, mean=3))
d$yy <- rnorm(100)*10 + d$x^2
d$zz <- d$y^2

plot_marg(d=d, x='xx', y='yy', marg='hist')
plot_marg(d=d, x='xx', y='yy', marg='hist', breaks=22)
plot_marg(d=d, x='xx', y='yy', marg='dens')
plot_marg(d=d, x='xx', y='yy', over='zz', marg='ortho')

}
