% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmarg.R
\name{plotmarg}
\alias{plotmarg}
\title{Scatterplot with marginals}
\usage{
plotmarg(d, x, y, over, marg = c("dens", "ortho"), ...)
}
\arguments{
\item{d}{array of data, where rows = SUs and cols = variables of interest}

\item{x}{character string for which column in \code{d} to place on x axis}

\item{y}{character string for which column in \code{d} to place on y axis}

\item{over}{character string for which column in \code{d} to overlay}

\item{marg}{either \code{'dens'} for density plots,
or \code{'ortho'} for orthogonal fitlines}
}
\value{
A ggplot2 object
}
\description{
Scatterplot with marginal densities or orthogonal fit lines, useful
    for visualizing gradients on ordinations.
}
\examples{
require(vegan)
data(dune, dune.env)
id <- cbind(scores(metaMDS(dune)), dune.env)
id$Moisture <- as.numeric(levels(id$Moisture))
plotmarg(d=id, x='NMDS1', y='NMDS2', over='Moisture', marg='dens')
plotmarg(d=id, x='NMDS1', y='NMDS2', over='Moisture', marg='ortho')

}
