% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_climatena.R
\name{prep_climatena}
\alias{prep_climatena}
\title{Prepare data for ClimateNA}
\usage{
prep_climatena(
  ID1,
  ID2 = NA,
  lat,
  long,
  el = NA,
  file = "./to_climatena.csv",
  ...
)
}
\arguments{
\item{ID1}{vector of primary identifier values (required)}

\item{ID2}{vector of secondary identifier values (optional)}

\item{lat}{latitude values, in decimal degrees (required)}

\item{long}{longitude values, in decimal degrees (required)}

\item{el}{vector of elevation values, in meters a.s.l. (optional)}

\item{file}{filename for .csv output}

\item{...}{further arguments currently ignored}
}
\value{
Writes a csv file ready for ClimateNA querying.
}
\description{
Prepare data to query climate values using ClimateNA (>= 5.0),
    with required formats and column names.
}
\details{
Works with ClimateNA (Wang et al. 2016) versions >= 5.0.
}
\examples{
data(braun)
env <- braun$env
tmp <- tempfile(fileext = '.csv')
prep_climatena(ID1=rownames(env), ID2=NA, lat=env$lat,
               long=env$lon, el=NA, file=tmp)
head(read.csv(tmp), 10)
unlink(tmp)

}
\references{
Wang, T., Hamann, A., Spittlehouse, D.L., Carroll, C. 2016. Locally
    downscaled and spatially customizable climate data for
    historical and future periods for North America. PLoS ONE
    11(6).
}
