% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_pcord.R
\name{r_pcord}
\alias{r_pcord}
\title{R to PC-ORD}
\usage{
r_pcord(x, filename = NULL, outdir = getwd(), header = TRUE, ...)
}
\arguments{
\item{x}{array of species or environmental data, where rows = SUs
and row.names = SU identifiers}

\item{filename}{string, name for new file; if NULL then returns an R object; if
file already exists then a timestamp is added}

\item{outdir}{string, directory to which file will be written}

\item{header}{logical, add a PC-ORD header?}

\item{...}{further arguments passed to other methods}
}
\value{
.xls file (optionally an R object) with automatic PC-ORD header
}
\description{
Export R data.frame for use in PC-ORD
}
\details{
Writes a file compatible with PC-ORD (McCune and Mefford 2011), and
    automates the addition of a PC-ORD header. Empty cells not
    allowed.  Column names longer than 10 characters are
    automatically abbreviated to \emph{unique} 10-char names;
    PC-ORD abbreviates but doesn't check for duplicates.

    Call \code{r_pcord(header = TRUE)}, then in PC-ORD do: File >
    Import > Main matrix > Excel spreadsheet (with PC-ORD header).
}
\examples{
# species abundance data
spe <- data.frame(matrix(rnorm(9, 10), 3, 3))
colnames(spe) <- c("Acer rubrum",
                   "Acer saccharum",
                   "Acer saccharinum")
lookin <- getwd()
r_pcord(spe, filename="spe_example", outdir=lookin)

# environmental data including some factors
env <- data.frame(soil = c("A","B","D"),
                  aspect = c("N","S","SE"),
                  slope = c(1,17,9))
r_pcord(env, filename="env_example", outdir=lookin)

# header not strictly required
r_pcord(env, filename="env_example", outdir=lookin, header=FALSE)

# allows no element to be empty or NA
env[2,3] <- NA
r_pcord(env, filename="env_example", outdir=lookin)

# allows only xls output
r_pcord(env, filename="env_example.csv", outdir=lookin)

}
\references{
McCune, B., and M. J. Mefford. 2011. PC-ORD. Multivariate Analysis
    of Ecological Data. Version 6. MjM Software Design, Gleneden
    Beach, OR.
}
