% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root mean square error (RMSE)}
\usage{
rmse(y, ypred, ...)
}
\arguments{
\item{y}{vector of reference values against which deviations
are measured}

\item{ypred}{vector of 'predicted' or 'sampled' values to evaluate,
of same length as \code{y}}

\item{...}{further arguments passed to other methods}
}
\value{
Numeric value, which is the root mean square error (RMSE).
}
\description{
Calculate root mean square error (RMSE) given
reference and predicted values
}
\details{
This is one way to assess prediction accuracy.
}
\examples{
n <- 99
y <- c(1:n)
ypred <- x + rnorm(n, 0.1)
rmse(y, ypred)

}
\seealso{
\link{https://en.wikipedia.org/wiki/Root-mean-square_deviation}
}
