% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Sample standard error}
\usage{
sem(x, ...)
}
\arguments{
\item{x}{vector of values to evaluate}
}
\value{
Numeric value which is the sample standard error
}
\description{
Calculate sample standard error for a vector.
}
\details{
Uses \emph{n-1} in denominator to correct for small-sample bias.
}
\examples{
x <- c(1:99)
sem(x)

}
\seealso{
\code{\link[base]{sd}}
}
