% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{asin_sqrt}
\alias{asin_sqrt}
\alias{genlogtrans}
\alias{normalize}
\title{Transforms}
\usage{
asin_sqrt(x, ...)

genlogtrans(x, ...)

normalize(x, ...)
}
\arguments{
\item{x}{vector or matrix of values to transform}

\item{...}{further arguments passed to other methods}
}
\value{
Values on the modified scale.
}
\description{
Transformations appropriate for ecological data.
}
\details{
\code{asin_sqrt} = Arc-sine square root transform:\cr
Appropriate for vegetation cover and other proportion data.  Forces
data to 0-1 scale if any original values were outside this range.
Follows pg. 69 of McCune & Grace (2002).

\code{genlogtrans} = Generalized log-transform:\cr
Log transform that properly handles zeros and very small values.
Preserves original order of magnitudes while accounting for values
= 0 and << 1. Follows pg. 69 of McCune & Grace (2002).

\code{normalize} = Normalize 0-1:\cr
Relativizes data to 0-1 scale based on min/max values.
}
\examples{
set.seed(21)
x  <- rnorm(99, 50, 20)
x1 <- asin_sqrt(x)
x2 <- genlogtrans(x)
x3 <- normalize(x)
par(mfrow=c(1,3))
plot(x, x1)
plot(x, x2)
plot(x, x3)
par(mfrow=c(1,1))

}
\references{
McCune, B., and J. B. Grace. 2002. Analysis of Ecological
    Communities. MjM Software, Gleneden Beach, Oregon, USA. 304 pp.
}
