% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{colvec}
\alias{colvec}
\alias{surfcol}
\title{Color utilities}
\usage{
colvec(x, n = 99, pal, zeroctr = FALSE, ...)

surfcol(x, ngrid, pal, ...)
}
\arguments{
\item{x}{vector of values to evaluate}

\item{n}{number of unique colors to be used in the palette}

\item{pal}{optional specification for a color palette; defaults to
\code{viridis::inferno} palette if missing}

\item{zeroctr}{logical, should the color scale diverge to be symmetrical
around zero?}

\item{...}{further arguments passed to other methods}

\item{ngrid}{number of grid cells in the matrix of values plotted
in \code{\link[graphics]{persp}}}
}
\value{
Vector or matrix of color values
}
\description{
From given values, generate colors for plotting points or surfaces.
}
\examples{
# generate data
x <- seq(-1.95, 1.95, length = 30)
y <- seq(-1.95, 1.95, length = 35)
z <- outer(x, y, function(a, b) a*b^2)

# color surface of perspective plot by 'z' value
ngrid <- prod(dim(z))
f     <- surfcol(z, ngrid=ngrid, alpha=0.8)
persp(x, y, z, col=f, phi=30, theta=-30)

# color points by numeric values
plot(x, y[1:30], col=colvec(x))

# color points by factor values
trmt <- gl(3, 10, labels = c('Q', 'R', 'S'))
plot(x, y[1:30], col=colvec(rev(trmt)))

# divergent colors centered on zero
set.seed(21)
x <- (rbeta(99, 0.5, 1) - 0.15) * 100
par(mfrow=c(1,3))
plot(x, col=colvec(x))
plot(x, col=colvec(x, zeroctr=TRUE))
plot(x, col=colvec(pseudo_log(x), zeroctr=TRUE))
par(mfrow=c(1,1))

}
