% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reshape.R
\name{utils_reshape}
\alias{utils_reshape}
\title{Reshape long to wide for perspective plots}
\usage{
utils_reshape(data, ...)
}
\arguments{
\item{data}{data.frame of 3 columns, of which first two are x,y
locations and third column is z values}

\item{...}{further arguments passed to other methods}
}
\value{
List with components x$x and x$y for x and y axes, and x$z is a
     matrix containing the values to be plotted
}
\description{
Safely reshape 'long' data.frame to 'wide' format suitable for
    'persp', 'image', 'contour', 'filled.contour', 'heatmap', etc.
}
\details{
Solves the dreaded "Error: increasing 'x' and 'y' values expected".
}
\examples{
set.seed(23)
d <- expand.grid(lon=seq(-125.5, -110.0, by=0.2),
                 lat=seq(40.5, 49.5, by=0.2))
d$z <- rnorm(nrow(d), 30, .2) * d$lat * d$lon
res <- reshape_p(d)
image(res)
contour(res, add=T, lwd=2, nlevels=12)
persp(res, theta=0, phi=45)
filled.contour(res, color.palette=heat.colors)
heatmap(res$z)

}
\seealso{
\code{\link[labdsv]{matrify}}
}
