% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{geom_mean}
\alias{geom_mean}
\alias{harm_mean}
\alias{sem}
\alias{cv}
\alias{rmse}
\alias{mae}
\title{Summary statistics utilities}
\usage{
geom_mean(x, na.rm = TRUE, zero.rm = FALSE, ...)

harm_mean(x, na.rm = TRUE, zero.rm = FALSE, ...)

sem(x, na.rm = TRUE, ...)

cv(x, na.rm = TRUE, ...)

rmse(x, y, na.rm = TRUE, ...)

mae(x, y, stdz = FALSE, na.rm = TRUE, ...)
}
\arguments{
\item{x}{vector of values to evaluate}

\item{na.rm}{logical indicating whether NA values in \code{x}
should be removed before proceeding}

\item{zero.rm}{logical indicating whether zeros in \code{x} should
be removed before calculating harmonic or geometric means}

\item{...}{further arguments passed to other methods}

\item{y}{vector of 'predicted' values to compare against \code{x}}

\item{stdz}{logical, standardize output by range of \code{x}?}
}
\value{
Numeric value.
}
\description{
Vectorized summary statistics, including geometric mean, harmonic
mean, sample standard error (SE), coefficient of variation (CV),
root mean square error (RMSE),  mean absolute error (MAE), and
sensitivity.
}
\details{
For vectors including at least one zero, results of
\code{geom_mean} and \code{harm_mean} are always 0 by
definition, unless \code{zero.rm=TRUE}.

Like \code{\link[stats]{sd}}, \code{sem} uses \emph{n-1} in
denominator to correct for small-sample bias.

\code{rmse} is one way to assess prediction accuracy.

\code{mae} gives a measure of sensitivity when \code{stdz=TRUE}.

These functions return NA when NAs present and \code{na.rm=TRUE}.
}
\examples{
# test data
xx <- c(-1, 0, 1, 4, 77, NA)

# harmonic mean
harm_mean(xx, na.rm=TRUE, zero.rm=FALSE)     # 0 by definition
harm_mean(xx, na.rm=TRUE, zero.rm=TRUE)      # 15.20988

# geometric mean
### NOT RUN:
# geom_mean(xx, na.rm=TRUE, zero.rm=FALSE))  # fails for neg vals
### END NOT RUN
xx <- xx[-1]                                 # remove negative values
geom_mean(xx, na.rm=TRUE, zero.rm=FALSE)     # 0 by definition
geom_mean(xx, na.rm=TRUE, zero.rm=TRUE)      # 6.753313

# standard error of the mean
sem(xx)                       # 21.76899

# coefficient of variation
cv(xx)                        # 183.9268

# root mean squared error
set.seed(23)
xx <- c(-1, 0, 1, 4, 77, NA)
yy <- xx+rnorm(length(xx), 10)
rmse(xx, yy)                  # 10.71919
rmse(yy, xx)                  # same, order invariant

# mean absolute error
mae(xx, yy, stdz=FALSE)       # 10.69236

# range-standardized mean absolute error (aka sensitivity)
mae(xx, yy, stdz=TRUE)        # 0.1370815
mae(yy, xx, stdz=TRUE)        # 0.135684 -- order matters!

}
\seealso{
\code{\link[stats]{sd}}
}
