% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_text.R
\name{clean_text}
\alias{clean_text}
\title{Clean character vector}
\usage{
clean_text(x, lower = F, ...)
}
\arguments{
\item{x}{character vector}

\item{lower}{logical, convert all characters to lowercase?}

\item{...}{further arguments passed to other methods}
}
\value{
Cleaned-up character vector of the same length and with the same
attributes as \code{x}.
}
\description{
Clean a character vector (text string) by replacing punctuation and
whitespaces with underscores.
}
\details{
Useful to clean a vector of species names, perhaps obtained from a
biodiversity database, to a machine-readable form.
}
\examples{
# sloppy
x <- c('Alnus___incana', 'Alnus incana ssp.tenuifolia',
       'Alnus rubra  ', '  Alnus viridis',
       'Alnus viridis ssp. crispa', 'Alnus viridis ssp.. sinuata')
# cleaned
clean_text(x)
clean_text(x, lower=TRUE)
clean_text('Alnus viridis ssp.. sinuata ')

}
