% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe}
\usage{
describe(x, na.rm = TRUE, digits = 2, type = 1, ...)
}
\arguments{
\item{x}{vector of values to be summarized}

\item{na.rm}{logical, defaults to TRUE}

\item{digits}{numeric, number of digits for rounding}

\item{type}{default=1, an integer between 1 and 3 selecting one of
the algorithms for computing skewness detailed in
\code{\link[e1071]{kurtosis}}}
}
\value{
List with 4 moments: \itemize{
  \item mean = mean
  \item sd = standard deviation
  \item var  = variance
  \item sem = standard error of the mean
  \item cv = coefficient of variation
  \item n = count of real elements
  \item NAs = count of NA elements
  \item skw = skewness
  \item krt = kurtosis
  }
}
\description{
Descriptive summary statistics
}
\details{
Method-of-moments summaries for first 4 moments.  \itemize{
  \item NEG kurtosis = flat distribution (platykurtic)
  \item POS kurtosis = peaked distribution (leptokurtic)
  \item ZERO kurtosis ~ the normal distribution (mesokurtic)
  \item NEG skewness indicates mean < median (left-skewed)
  \item POS skewness indicates mean > median (right-skewed)
  }
}
\examples{
# require(e1071)
n <- 99
x <- c(rnorm(n, 0.1), NA)
describe(x)

}
\seealso{
\code{\link[ecole]{mom}},
\code{\link[e1071]{kurtosis}},
\code{\link[e1071]{skewness}}
}
