% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mx_diversity.R
\name{mx_diversity}
\alias{mx_diversity}
\title{Diversity of a species matrix}
\usage{
mx_diversity(x, ...)
}
\arguments{
\item{x}{array of species data, where rows = SUs and cols = species}
}
\value{
Dataframe with 8 numeric values: \itemize{
  \item gamma = count of species in entire matrix
  \item alpha = average count of species per SU
  \item beta  = Whittaker's beta diversity = [(gamma/alpha) - 1]
  \item halfchanges = beta diversity by exponential transformation
            of average Sorenson dissimilarities (pg 31 of McCune &
            Grace 2002)
  \item dbi   = Dust Bunny Index of McCune and Root (2015)
  \item prop0 = proportion of zero elements (McCune and Root 2015)
  \item propnoshare = proportion of no-share SU pairs
  \item N     = number of SUs
  }
}
\description{
Calculate diversity properties of a species matrix:
    Gamma, Alpha, Beta, Halfchanges, Dust Bunny Index, others.
}
\details{
Calculates eight measures of diversity in the matrix. For an
    example applied to many simulated datasets, see Smith (2017).
}
\examples{
# species abundance data
set.seed(1917)
spe <- data.frame(matrix(rnorm(30, 10, 50), 10, 3))
spe[spe<0] <- 0
colnames(spe) <- c('Acer rubrum','Acer saccharum','Acer saccharinum')
spe
mx_diversity(spe)

}
\references{
McCune, B., and J. B. Grace. 2002. Analysis of Ecological
    Communities. MjM Software, Gleneden Beach, Oregon, USA. 304 pp.

McCune, B., and H.T. Root. 2015. Origin of the dust bunny
    distribution in ecological community data. Plant Ecology
    216(5): 645-656.

Smith, R.J. 2017. Solutions for loss of information in
    high-beta-diversity community data. Methods in Ecology and
    Evolution 8(1): 68-74.
}
