% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ord_bc.R
\name{ord_bc}
\alias{ord_bc}
\title{Bray-Curtis ordination, aka polar ordination}
\usage{
ord_bc(
  x,
  method = "bray",
  endpoint = c("var.reg", "BC.original", "PC_ORD.original"),
  ends = c(NA, NA),
  ...
)
}
\arguments{
\item{x}{array, where rows = SUs and cols = variables such as environmental
or traits values.}

\item{method}{dissimilarity index, passed to `vegan::vegdist`.}

\item{endpoint}{Endpoint selection method. `'BC.original'` is Bray-Curtis
original of selecting most-distant SUs.  `'PC_ORD.original'` is the
deprecated method from PC-ORD, whereby the first endpoint has the highest
sum of distances from other sites, and the second endpoint has the highest
distance from the first endpoint. `'var.reg'` is the variance regression
method now default in PC-ORD (McCune and Mefford 2017).}

\item{ends}{Numeric, index two sample units as endpoints on first axis.}

\item{...}{further arguments not currently passed to other functions.}
}
\value{
List containing items:\cr
- \code{SSresid}:  sums-of-squares of residual distances per axis.\cr
- \code{varexplained}:  variance explained per axis.\cr
- \code{endpoints}:  endpoint sample units per axis.\cr
- \code{scores}:  ordination scores.\cr
}
\description{
Bray-Curtis ordination.
}
\details{
Bray-Curtis ordination. See references.
}
\examples{
data(smoky, package='ecole')
spe <- smoky$spe
o1 <- ord_bc(spe, method='bray', endpoint='BC.original')
o2 <- ord_bc(spe, method='bray', endpoint='BC.original', ends=c(1,12))
ecole::set_par(2)
plot(scores(o1), type='n')
text(scores(o1))
plot(scores(o2), type='n')
text(scores(o2))

}
\references{
Beals, E.W. 1984. Bray-Curtis ordination: an effective strategy for analysis
of multivariate ecological data. Advances in Ecological Research 14: 1-55.\cr

Bray, J.R., and J.T. Curtis. 1957. An ordination of the upland forest
communities of southern Wisconsin. Ecological Monographs 27: 325-349.\cr

McCune, B. and E.W. Beals. 1993. History of the development of Bray-Curtis
ordination. Pgs 67-80 in J.S. Fralish, R.P. McIntosh, and O.L. Loucks (eds).
John T. Curtis: Fifty Years of Wisconsin Plant Ecology, Wisconsin Academy of
Science, Arts, and Letters, Madison, WI.\cr

McCune, B., and M.J. Mefford. 2017. PC-ORD. Multivariate Analysis of
Ecological Data. Version 7. MjM Software Design, Gleneden Beach, OR.\cr
}
