% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plo.R
\name{plo}
\alias{plo}
\title{Simple scatterplot with useful defaults}
\usage{
plo(x, y = NULL, col = "#00000020", cex = 0.5, pch = 16, ...)
}
\arguments{
\item{x, y}{vectors of x- and y-axis values.}

\item{col}{vectors of point colors.}

\item{cex}{numeric vector for scaling point size.}

\item{...}{further arguments passed to \code{graphics::plot.default}}
}
\value{
A plot object.
}
\description{
Scatterplot with sensible default settings that permit
    overplotting many points.
}
\details{
Useful default settings for a simple scatterplot, closed circles
    with transparent black color that permits overplotting.
    Follows \code{graphics::plot.default}.
}
\examples{
set.seed(23)
x <- runif(9999)
y <- rnorm(9999)
plo(x, y, las=2, bty='l')

}
