% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of species abundance matrix}
\usage{
plot_heatmap(
  x,
  xord = "none",
  yord = "none",
  logbase = FALSE,
  labcex = 0.7,
  col,
  xexp = 1,
  yexp = 1,
  ...
)
}
\arguments{
\item{x}{array of species data, where rows = SUs and cols = species}

\item{xord}{logical, order columns by species weighted averages?}

\item{yord}{logical, order rows by species (unweighted) averages?}

\item{logbase}{numeric, logarithm base to scale plot contrast}

\item{labcex}{numeric, label expansion factor; defaults to 0.7}

\item{col}{vector defining a color palette}

\item{xexp, yexp}{expansion factor making room for x,y labels}

\item{...}{further arguments passed to other methods}
}
\value{
None, plots to device.
}
\description{
Plot site x species matrix with heatmap colors and proper labels.
}
\details{
Plots a heatmap to quickly visualize species abundances/occurences.
}
\examples{
data(smoky)
z <- smoky$spe
# basic
plot_heatmap(z)

# pretty
r <- colorRampPalette(c('transparent','darkgreen'))(99)
plot_heatmap(z, yexp=1.7, logbase=10, asp=1, col=r)

# various ways to order
plot_heatmap(z, xord='wa', yord='none')
plot_heatmap(z, xord='mean', yord='none')
plot_heatmap(z, xord='wa', yord='wa')

}
\seealso{
\code{\link[vegan]{tabasco}} for comparison.
}
