% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joy.R
\name{plot_joy}
\alias{plot_joy}
\title{Joy plots}
\usage{
plot_joy(m, w = NULL, scaled = TRUE, excl = 0, fcol = "#00000080",
  lcol = "#000000", lwid = 1, xinc = 0, xexp = 0.2,
  yinc = ncol(m), yexp = 1.2, ypad = 1, pts = NULL, labcex = 0.5,
  ...)
}
\arguments{
\item{m}{matrix or data.frame, where each column vector will have
its own density polygon}

\item{w}{matrix or data.frame, of same dimensions as \code{m}, with
non-negative weights (in ecology, this might be species
abundances)}

\item{scaled}{logical, should polygons be scaled to maximum?}

\item{excl}{numeric, value(s) to exclude from density estimate}

\item{fcol}{polygon fill color(s)}

\item{lcol}{polygon line color}

\item{lwid}{polygon line width}

\item{xinc}{horizontal offset for each polygon (0 = aligned)}

\item{xexp}{x-axis expansion factor for entire plot}

\item{yinc}{vertical offset for ea polygon}

\item{yexp}{vertical expansion factor for each polygon}

\item{ypad}{vertical padding for entire plot}

\item{pts}{optional vector of points to plot onto each row}

\item{labcex}{expansion factor for y-axis text labels}

\item{...}{further arguments passed to plot function}
}
\value{
None, plots to device.
}
\description{
Plot small multiples of many overlapping density distributions,
    also known as a joy plot or stacked distributions.
}
\details{
Plots one density polygon for each column in the data.frame or
    matrix, resulting in potentially overlapping small multiples.
    Polygons can be scaled vertically to their maximum, or else
    allowed to vary freely.  Setting \code{yexp = 1},
    \code{ypad = 1} and \code{scaled = T} gives scaled polygons
    that fill each row with no overlap.

NA values are omitted prior to density calculations, and other
    special values may also be omitted by, for example,
    \code{excl = 0} or \code{excl = c(0,1,2)}.

Weights can be useful for species abundance-weighted density
    estimation.
}
\examples{
# synthetic example
set.seed(23)
x <- rnorm(99,10,1)
m <- data.frame(x = x,
                y = x^1.2 + rnorm(99,0,1),
                z = x^1.4 + rnorm(99,0,1))
plot_joy(m, ypad=1.05, labcex=1)        # allow overlap
plot_joy(m, yexp=1, ypad=1, labcex=1)   # forbid overlap
plot_joy(m, yexp=1, ypad=1, labcex=1, fcol=4:6) # color vector

# real example
data(braun)
abu <- braun$spe              # species abundances
abu <- abu[colMeans(abu)>1]   # keep only most dominant species
env <- braun$env$bio1         # environment
occ <- abu
occ[occ>0] <- 1               # binary occurrences matrix
y   <- occ*env                # matrix of env values at occurrences
lab <- ' Annual Mean Temperature'
set_par(2)
plot_joy(y, w=NULL, xlab=lab) # unweighted
plot_joy(y, w=abu,  xlab=lab) # abundance-weighted

}
