% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loess.R
\name{plot_loess}
\alias{plot_loess}
\alias{plot_lm}
\title{Plotting locally-weighted or linear regression line}
\usage{
plot_loess(x, y, col = "#00000040", lcol = "#FF0000BF", cex = 0.8,
  pch = 16, las = 1, bty = "L", xlab = NULL, ylab = NULL, ...)

plot_lm(x, y, col = "#00000040", lcol = "#FF0000BF", cex = 0.8,
  pch = 16, las = 1, bty = "L", xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{x, y}{vectors of x- and y-axis values.}

\item{col}{vector of point colors.}

\item{lcol}{vector of line colors.}

\item{cex}{numeric vector for scaling point size.}

\item{pch}{a vector of plotting characters or symbols: see
\code{\link[graphics]{points}}.}

\item{las}{numeric, style of axis labels: see \code{\link[graphics]{par}}.}

\item{bty}{character string, determines type of box drawn around plot: see
\code{\link[graphics]{par}}.}

\item{xlab, ylab}{character strings for axis labels.}

\item{...}{further arguments passed to functions.}
}
\value{
A plot object.
}
\description{
Fit a linear, or else locally-fitted polynomial surface determined
   by a numerical predictor.
}
\details{
Useful default settings for a scatterplot with locally-weighted
    regression line.
}
\examples{
set.seed(23)
x <- sin(1:55/pi)
y <- rnorm(55, 0, 0.45)+x
set_par(3)
plot_loess(x, y, span=0.10, main='Span = 0.10')
plot_loess(x, y, span=0.20, main='Span = 0.20')
plot_loess(x, y, span=0.40, main='Span = 0.40')

}
\seealso{
\code{\link[stats]{loess}}
}
