% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ortho.R
\name{plot_ortho}
\alias{plot_ortho}
\title{Scatterplot with orthogonal fit line}
\usage{
plot_ortho(x, y, dist = FALSE, col, ann = TRUE, ...)
}
\arguments{
\item{x}{vector of x-axis values}

\item{y}{vector of y-axis values}

\item{dist}{logical, convert values to interpoint Euclidean distances?}

\item{col}{single value or vector of point colors}

\item{ann}{logical, should annotations be added to top of plot?}

\item{...}{further arguments passed to other methods}
}
\value{
A plot object
}
\description{
Scatterplot with orthogonal fit line and annotations.  Also
permits converting from points to interpoint Euclidean distances.
}
\details{
Uses \code{\link[stats]{prcomp}} to fit orthogonal line. Data are
first centered but not scaled for PCA step, therefore, consider
carefully whether unscaled data are meaningful in your use case.
Plot aspect is fixed at \code{asp=1}.

Converting values to interpoint distances may be useful for
comparing interpoint distances in e.g., ordination scores.
}
\examples{
set.seed(23)
N  <- 99
x  <- runif(N,0,1)
y1 <- x + rnorm(N, 0, .1)
y2 <- x + rnorm(N, 0, .5)
y3 <- rnorm(N, 0, x*.3)
op <- par(mfrow=c(2,2))
plot_ortho(x, y1)
plot_ortho(x, y2)
plot_ortho(x, y3)
plot_ortho(y3, x)
par(op)

}
