% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_color.R
\name{colvec}
\alias{colvec}
\alias{surfcol}
\title{Color utilities}
\usage{
colvec(x, n = 99, alpha = 0.6, begin = 0.2, end = 0.9, dir = 1,
  pal, ...)

surfcol(x, ngrid, alpha = 0.6, begin = 0.2, end = 0.9, dir = 1,
  pal, ...)
}
\arguments{
\item{x}{vector of values to evaluate}

\item{n}{number of unique colors to be used in the palette}

\item{alpha}{alpha transparency in [0,1]}

\item{begin}{hue in [0,1] at which the viridis colormap begins}

\item{end}{hue in [0,1] at which the viridis colormap ends}

\item{dir}{order of colors in the scale; if 1, the default, colors
are ordered from darkest to lightest; if -1, the order of colors
is reversed}

\item{pal}{optional specification for a color palette; defaults to
\code{viridis::inferno} palette if missing}

\item{...}{further arguments passed to other methods}

\item{ngrid}{number of grid cells in the matrix of values plotted
in \code{\link[graphics]{persp}}}
}
\value{
Vector or matrix of color values
}
\description{
From given values, generate colors for plotting points or surfaces.
}
\examples{
# generate data
x <- seq(-1.95, 1.95, length = 30)
y <- seq(-1.95, 1.95, length = 35)
z <- outer(x, y, function(a, b) a*b^2)

# color surface of perspective plot by 'z' value
ngrid <- prod(dim(z))
f <- surfcol(z, ngrid=ngrid, alpha=0.8)
persp(x, y, z, col=f, phi=30, theta=-30)

# color points by numeric value
plot(x, y[1:30], col=colvec(x), pch=16)

# color points by factor value
trmt <- gl(3, 10, labels = c('Q', 'R', 'S'))
plot(x, y[1:30], col=colvec(rev(trmt)), pch=16)

}
