% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{set_par}
\alias{set_par}
\alias{add_label}
\alias{add_text}
\title{Plotting utilities}
\usage{
set_par(panels = 1, pty = "s", ...)

add_label(char = "a", bold = FALSE, ...)

add_text(x, y, labels, bold = FALSE, ...)
}
\arguments{
\item{panels}{number of panels (1,2,4) to plot.}

\item{pty}{a character specifying the type of plot region to be
used; "s" generates a square plotting region and "m" generates
the maximal plotting region.}

\item{...}{further arguments passed to other methods (i.e.,
\code{\link[graphics]{par}})}

\item{char}{alphabetic character to add as panel label.}

\item{bold}{logical, should panel label be bold?}

\item{x}{numeric, \emph{relative} horizontal position}

\item{y}{numeric, \emph{relative} vertical position}

\item{labels}{character vector of text to plot}
}
\value{
`set_par` sets graphical parameters; `add_label` adds panel label;
    `add_text` adds text to the plotting area.
}
\description{
Set "neat" graphical parameters suitable for publication, and also
    label panels on multi-panel figures.
}
\examples{
N  <- 20
op <- set_par(4, pch=16)
plot(1:N, N:1)
add_label('a')
plot(sin(1:N), cos(N:1))
add_label('b')
plot(1:N, 1:N+rnorm(N,0,1), xlim=c(0,2*N), xlab='x', ylab='f(x)')
add_label('c')
plot(1:N, 1:N+rnorm(N,0,5), ylim=c(0,2*N), xlab='x', ylab='f(x)')
add_label('d')
add_text(.1, .9, '+item1')
add_text(.5, .5, '+item2')
add_text(.9, .1, '+item3') # extends beyond plot area
par(op)

}
