% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_stats.R
\name{geom_mean}
\alias{geom_mean}
\alias{harm_mean}
\alias{sem}
\alias{cv}
\alias{rmse}
\title{Summary statistics utilities}
\usage{
geom_mean(x, na.rm = TRUE, zero.rm = FALSE, ...)

harm_mean(x, na.rm = TRUE, zero.rm = FALSE, ...)

sem(x, na.rm = TRUE, ...)

cv(x, na.rm = TRUE, ...)

rmse(x, xpred, na.rm = TRUE, ...)
}
\arguments{
\item{x}{vector of values to evaluate}

\item{na.rm}{logical indicating whether NA values in x should be
removed before proceeding}

\item{zero.rm}{logical indicating whether zeros in x should be
removed before calculating harmonic or geometric means}

\item{...}{further arguments passed to other methods}

\item{xpred}{vector of 'predicted' values to compare against x}
}
\value{
Numeric value
}
\description{
Vectorized summary statistics, including geometric mean, harmonic
mean, sample standard error (SE), coefficient of variation (CV),
and root mean square error (RMSE).
}
\details{
Results of \code{geom_mean} and \code{harm_mean} are always 0 by
definition, unless \code{zero.rm=TRUE}.

Like \code{\link[stats]{sd}}, \code{sem} uses \emph{n-1} in
denominator to correct for small-sample bias.

\code{rmse} is one way to assess prediction accuracy.

These functions return NA when NAs present and \code{na.rm=TRUE}.
}
\examples{
# test data
xx <- c(-1, 0, 1, 4, 77, NA)

# harmonic mean
harm_mean(xx, na.rm=T, zero.rm=F)  # 0 by definition
harm_mean(xx, na.rm=T, zero.rm=T)  # 15.20988

# geometric mean
geom_mean(xx, na.rm=T, zero.rm=F)  # should fail for negative values
xx <- xx[-1]                       # remove negative values
geom_mean(xx, na.rm=T, zero.rm=F)  # 0 by definition
geom_mean(xx, na.rm=T, zero.rm=T)  # 6.753313

# standard error of the mean
sem(xx)                            # 21.76899

# coefficient of variation
cv(xx)                             # 183.9268

# root mean squared error
set.seed(23)
rmse(xx, xx+rnorm(length(xx)), na.rm=T) # 1.033994

}
\seealso{
\code{\link[base]{sd}}
}
